Fix raster structure.
Change ref_clk and raster_clk from int to long, this allow to use this
structure even in 8bit microcontrollers where int is 16 bit long.

+++ SDK/lib/api/lcd_def.h
@@ -51,11 +51,11 @@ typedef struct
 {
 	unsigned int X;
 	unsigned int Y;
-	unsigned int ref_clk;
+	unsigned long ref_clk;
 	unsigned int pll_m;
 	unsigned int pll_n;
 	unsigned int pll_m2;
-	unsigned int raster_clk;
+	unsigned long raster_clk;
 	unsigned int palete_len;
 	unsigned int bus_size;
 	unsigned int hsw;
